src/utils.py
import numpy as np
from scipy.fft import fft  # Kept for potential future signal analysis
import serial
import RPi.GPIO as GPIO
import time

# Setup for mmWave UART (e.g., RD-03D)
UART_PORT = '/dev/ttyAMA0'  # Or '/dev/ttyS0' on Pi 5
BAUDRATE = 256000  # From mmWave docs
PHOTODIODE_PIN = 22  # For LiDAR pulse detection
GPIO.setup(PHOTODIODE_PIN, GPIO.IN, pull_up_down=GPIO.PUD_UP)  # Pull-up for stability

def detect_signal(duration=0.1):
    """Detect AV sensors: mmWave for radar/motion, photodiode for LiDAR pulses"""
    detected = False

    # mmWave radar detection (proxy for AV radar)
    try:
        ser = serial.Serial(UART_PORT, BAUDRATE, timeout=0.1)
        ser.write(b'\xFF\xFF\xFF\xFF')  # Wakeup/init command if needed (check module docs)
        data = ser.read(100)  # Read raw bytes
        ser.close()
        # Parse for motion/presence (simplified; use module library if available)
        if b'motion' in data or len(data) > 10:  # Placeholder parse; refine with actual protocol
            detected = True
    except Exception as e:
        print(f"mmWave error: {e}")

    # LiDAR pulse detection (IR photodiode)
    start_time = time.time()
    while time.time() - start_time < duration:
        if GPIO.input(PHOTODIODE_PIN) == GPIO.LOW:  # Pulse detected (active low with amp)
            detected = True
            break
        time.sleep(0.001)  # Poll rate

    return detected